Move EndOfStudyView from MTB app to shared framework
