+++ SwiftPackage/Sources/BridgeClientUI/Resources/Images.xcassets/End of Study View/Contents.json
@@ -1,0 +1,6 @@
+{
+  "info" : {
+    "author" : "xcode",
+    "version" : 1
+  }
+}
