+++ SwiftPackage/Sources/BridgeClientUI/Resources/Images.xcassets/End of Study View/End of Study.imageset/Contents.json
@@ -1,0 +1,12 @@
+{
+  "images" : [
+    {
+      "filename" : "Group 4346.svg",
+      "idiom" : "universal"
+    }
+  ],
+  "info" : {
+    "author" : "xcode",
+    "version" : 1
+  }
+}
