+++ SwiftPackage/Sources/BridgeClientUI/Resources/Images.xcassets/End of Study View/Uploading.imageset/Contents.json
@@ -1,0 +1,12 @@
+{
+  "images" : [
+    {
+      "filename" : "Group 4017.svg",
+      "idiom" : "universal"
+    }
+  ],
+  "info" : {
+    "author" : "xcode",
+    "version" : 1
+  }
+}
